% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/only_hour.R
\name{only_hour}
\alias{only_hour}
\title{Import a dataset}
\usage{
only_hour(x)
}
\arguments{
\item{x}{Numeric variable}
}
\value{
Factor variable
}
\description{
More detailed description.
if there is any datetime column in the dataset and a step that you want to grab only the hour from the datatime column this package is useful to you.
}
\examples{
bike <- read.csv('bikeshare.csv')
head(bike$datetime)

only-hour<-sapply(bike$datetime,only_hour)

}
